Fix(users) : BugFixing Login and Register
