+++ dto/loginResponse.go
@@ -1,5 +1,7 @@
 package dto
 
+import "go-pos-api/domain"
+
 type LoginResponse struct {
 	ID       int    `json:"user_id"`
 	Name     string `json:"name"`
@@ -7,4 +9,16 @@ type LoginResponse struct {
 	Role     int    `json:"role"`
 	Merchant int    `json:"merchant"`
 	Token    string `json:"token,omitempty"`
+}
+
+func UserInput(user domain.User, token string) LoginResponse {
+	request := LoginResponse{
+		ID:       user.ID,
+		Name:     user.Name,
+		Email:    user.Email,
+		Role:     user.Role,
+		Merchant: user.Merchant,
+		Token:    token,
+	}
+	return request
 }
